First Commit
canaryio/canary is the spiritual successor to canaryio/sensord and
canaryio/canaryd.

It is like ping, but for URLs.

This initial commit introduces a lightweight command line tool,
`canary`, that allows you to measure a single URL.

It is composed of two main components: a sampler and a reporter.
Currently there is only a single sampler, one that relies on libcurl.

There are two reporters: TSV and logfmt, both sending structured data to
stdout.

I don't expect this API to remain stable, but it is usable.