Initial Librato Support
This changeset introduces an initial Librato Publisher, as well as the
supporting infrastructure to make such a thing possible (and to make it
easier to add others in the future).

In the pass, we re-arrange some of the core infra like so:

* we move the StdoutPublisher into its own package
* we move the TransportSampler into its own package
* we introduce the concept of canary.StatusCodeError to represent
  explicit error conditions in which the HTTP status code is out of
  range
* canary.Version is introduced and injected into HTTP User-Agent headers
* test coverage is increased
* golint passes

We introduce librato support via two packages, libratopublisher and
libratoaggregator.

`libratoaggregator` is a general purpose package that makes it easy to
aggregate metrics and flush them to the Librato API on a given
interval (currently hard coded to 5 seconds).

`libratopublisher` implements the `canary.Publisher` interface, and
takes care of gluing canary-proper to the `libratoaggregator`.

The metrics and configuration details are noted in the `canaryd` `README`.