Honor timeout in connect, always capture IP of remote host
Sorry for combining these into a single commit.  Got a little carried
away.

This ensures that connections to the remote host honor the timeout.
Previously the timeout only applied after the connection had been
made.

This also performs the IP address resolution outside of the connection
initiation, so that the DNS lookup time can be captured, as well.
Furthermore, the IP address is now *always* captured, not just after a
successful connection.